# openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SmartUnderwrite.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/Decision/{applicationId}/evaluate": {
      "post": {
        "tags": [
          "Decision"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Decision/{applicationId}/manual-decision": {
      "post": {
        "tags": [
          "Decision"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDecisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDecisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ManualDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Decision/{applicationId}/latest": {
      "get": {
        "tags": [
          "Decision"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Decision/{applicationId}/history": {
      "get": {
        "tags": [
          "Decision"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "loanApplicationId": {
            "type": "integer",
            "format": "int32"
          },
          "outcome": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "decidedByUserId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "decidedByUserName": {
            "type": "string",
            "nullable": true
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isAutomated": {
            "type": "boolean"
          }
        }
      },
      "DecisionOutcome": {
        "type": "integer"
      },
...
      "ManualDecisionRequest": {
        "required": [
          "outcome",
          "reasons"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "reasons": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "justification": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}